worker/requirements.txt
pymongo==4.7.3
redis==5.0.8
python-dotenv==1.0.1
